backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
